Next Auth - Proteccion de rutas

+++ 03-admin-todo/src/app/api/todo/route.ts
@@ -1,3 +1,4 @@
+import { getUserServerSession } from '@/auth/auth-actions'
 import prisma from '@/lib/prisma'
 import { NextResponse, NextRequest } from 'next/server'
 import { boolean, object, string } from 'yup'
@@ -26,10 +27,14 @@ const postSquema = object({
 
 export async function POST(request: NextRequest) {
 
+  const user = await getUserServerSession()
+
+  if ( !user ) return NextResponse.json("No autorizado", { status: 401 })
+
   try {
     const { complete, description } =  await postSquema.validate( await request.json() )  
     const todo = await prisma.todo.create({
-      data: { complete, description }
+      data: { complete, description, userId: user.id  }
     })
     return NextResponse.json(todo)
   } catch (error) {
@@ -39,9 +44,13 @@ export async function POST(request: NextRequest) {
 
 
 export async function DELETE( request: Request ) {
-  await prisma.todo.deleteMany({ where: { complete: true } })
+
+  const user = await getUserServerSession()
+
+  if ( !user ) return NextResponse.json("No autorizado", { status: 401 })
 
   try {
+    await prisma.todo.deleteMany({ where: { complete: true, userId: user.id  } })
       return NextResponse.json({ message: 'Todos borradas' })
   } catch (error) {
       return NextResponse.json(error, { status: 400 } )
